build: Consider SOURCE_DATE_EPOCH
The build is not reproducible, because the build date is compiled in.
Avoid that, if env variable SOURCE_DATE_EPOCH is set (like some build
environments do).

The approach used here is directly taken from the reproducible builds
project. See:
https://reproducible-builds.org/docs/source-date-epoch/#makefile

Console output before:

    AT91Bootstrap 3.8.12 (Mon Feb  4 13:02:06 CET 2019)

Console output after:

    AT91Bootstrap 3.8.12 (2019-01-01)

If SOURCE_DATE_EPOCH is not set, the old behaviour is kept.

Signed-off-by: Alexander Dahl <post@lespocky.de>