kconfig: add CPU_HAS_SCLK_BYPASS and fixed menuconfig
Some SoCs do not have the option to bypass the driving of the external
slow clock.
Namely, SAMA5D2 and SAMA5D4 do not have it.
Added a new SoC feature named CPU_HAS_SCLK_BYPASS which is by default
for all the other supported SoCs.
The menuconfig option for slow clock drive bypass (use an external oscillator,
not crystal) is now dependant on this SoC feature.
Rewrote the help text for better understanding of this feature.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Acked-by: Nicolas Ferre <nicolas.ferre@microchip.com>