board: sama5d27_som1_ek: Fix LED switch on
The code did set the output of pin PA27 to low.  According to SAMA5D27
SOM1 Kit1 User's Guide that is used as enable signal for an USB power
switch.  The red LED is actually connected active high to PA10 through a
MOSFET.  (Maybe that changed from an early prototype?)

If you power up the board without any bootloader, it goes to SAM-BA mode
and the red LED is on.  Switching that red LED on again in bootstrap
would give no benefit.  To see if the bootstrap was actually loaded,
this is changed to switch the green LED on now.

Fixes: a4f6043cfbb04b04d29abddb77658fbf6e66bfa0
Signed-off-by: Alexander Dahl <ada@thorsis.com>