driver: sdhc: fix BUSTEST_W/R commands
When sending the data for the bustest_w command, something happens
wrong and sending another command too fast makes the controller return errors.
To avoid it, reworked the data sending function to check for write ready
first.
When Data timeout occurs, it's also much better to reset the DAT and CMD lines
otherwise we will have errors later.
Reworked the sw workaround for missing TRFC to cope with the timeout situation
as well.
Retyped some variables as short, to be in sync with the real size of the
SDMMC register.
Because of this size mismatch, it would have been possible to write more than
needed to the hardware in ARM mode (as opposed to THUMB where it worked).

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>