Config: MEM_BANK and MEM_SIZE are not kernel-loading dependant
We need to be able to set MEM_BANK , MEM_BANK2 and MEM_SIZE
regardless of the fact that we load the Linux Kernel directly or not.
The size of the ext RAM and start of ext RAM can be used by other drivers.
For example, trustzone requires this information.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>