README: update windows instructions to use python3
The following section in README.txt needs to be updated.
installing
python3 vs python2, the symlinks aren't required anymore, it seems.

      $ pacman -S bc gawk gcc git make python2 tar
      $ cd /usr/bin
      $ ln -s python2.exe python.exe && ln -s python2-config
python-config

Further comments about this change: considering that macOS 10.8 to
10.15 come with Python 2.7 preinstalled (as /usr/bin/python2.7 and
/usr/bin/python), out-of-the-box user experience will be affected on
this OS. Not a big deal for CLI users, but for IDE users (MPLAB),
maybe we'll consider modifying the Makefile so that it tests for
python3 availability, and fallbacks to python2 or python, when it calls
scripts/addpmecchead.py (bypassing the shebang).

Because the 'python' package, on Windows with Minimal SYStem 2,
provides Python 3.7 as /usr/bin/python3 (and as /usr/bin/python too).

macOS 10.13 with MacPorts can achieve the same:
     port install python38
     port select --set python3 python38
     port select --set python python38

Debian also offers the flexibility to select the default Python version:
     update-alternatives --install /usr/bin/python python /usr/bin/python3.8

And Fedora too:
     alternatives --install /usr/bin/python python /usr/bin/python3.8

Suggested-by: Aubin Constans <aubin.constans@microchip.com>
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>